Updated for Rust 1.0

+++ examples/main.rs
@@ -1,16 +1,15 @@
-#![feature(phase)]
-#[phase(plugin)]
+#[macro_use]
 extern crate data_macro;
 
 data!(
     SomeData {
-        a: uint = 1             // fields can be set as normal
-        b: uint = a + 2         // fields can reference other fields
-        c: uint = mul_by_four(b)// fields can be set via functions
+        a: usize = 1             // fields can be set as normal
+        b: usize = a + 2         // fields can reference other fields
+        c: usize = mul_by_four(b)// fields can be set via functions
     }                           // commas and semicolons are unnecessary
 );
 
-fn mul_by_four(num: uint) -> uint {
+fn mul_by_four(num: usize) -> usize {
     println!("Multiplying {} by 4", num);
     num * 4
 }
